requirements.txt
einops==0.8.1
entmax==1.1
flops_profiler==0.1.2
heartpy==1.2.7
joblib==1.2.0
keras==2.10.0
numpy==1.23.5
lightning==2.0.1.post0
opencv_python==4.7.0.72
opencv_python_headless==4.6.0.66
pandas==1.4.4
PyYAML==6.0.2
scipy==1.9.1
timm==0.8.19.dev0
tqdm==4.65.0
tensorflow==2.10.0
retina-face==0.0.13
odfpy
openpyxl
wandb==0.15.1 
pydantic==1.10.9
scikit-image==0.24.0

